refactor: relocate insight modules to core
