+++ alpha_factory_v1/core/agents/__init__.py
@@ -3,5 +3,6 @@
 
 from .meta_refinement_agent import MetaRefinementAgent
 from .self_improver_agent import SelfImproverAgent
+from .base_agent import BaseAgent
 
-__all__ = ["MetaRefinementAgent", "SelfImproverAgent"]
+__all__ = ["MetaRefinementAgent", "SelfImproverAgent", "BaseAgent"]
