+++ alpha_factory_v1/core/scheduler/__init__.py
@@ -13,7 +13,7 @@ import random
 from rocketry import Rocketry
 from rocketry.conds import every
 
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src import self_improver
+from alpha_factory_v1.core.self_evolution import self_improver
 from alpha_factory_v1.core.monitoring import metrics
