+++ alpha_factory_v1/core/simulation/replay.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import List, Any, cast
 
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src.simulation import forecast, sector
+from alpha_factory_v1.core.simulation import forecast, sector
 
 
 BASE_DIR = Path(__file__).resolve().parents[2] / "data" / "historic_scenarios"
