+++ alpha_factory_v1/demos/alpha_agi_insight_v1/src/agents/adk_summariser_agent.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from .base_agent import BaseAgent
 from ..utils import messaging
 from ..utils.logging import Ledger
-from ..utils.tracing import span
+from alpha_factory_v1.core.utils.tracing import span
 
 
 class ADKSummariserAgent(BaseAgent):
