+++ alpha_factory_v1/demos/alpha_agi_insight_v1/src/agents/chaos_agent.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 from .base_agent import BaseAgent
 from ..utils import messaging
 from ..utils.logging import Ledger
-from ..utils.tracing import span
+from alpha_factory_v1.core.utils.tracing import span
 
 
 class ChaosAgent(BaseAgent):
