+++ alpha_factory_v1/demos/alpha_agi_insight_v1/src/agents/memory_agent.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 from .base_agent import BaseAgent
 from ..utils import messaging, logging as insight_logging
 from ..utils.logging import Ledger
-from ..utils.tracing import span
+from alpha_factory_v1.core.utils.tracing import span
 import os
 import json
 from pathlib import Path
