+++ alpha_factory_v1/demos/alpha_agi_insight_v1/src/agents/research_agent.py
@@ -15,7 +15,7 @@ from ..simulation import forecast, sector
 from ..utils import messaging, logging as insight_logging
 from ..utils.logging import Ledger
 from ..utils.retry import with_retry
-from ..utils.tracing import span
+from alpha_factory_v1.core.utils.tracing import span
 from typing import Callable, Awaitable, cast
 
 log = insight_logging.logging.getLogger(__name__)
