+++ alpha_factory_v1/demos/alpha_agi_insight_v1/src/agents/strategy_agent.py
@@ -12,7 +12,7 @@ from .base_agent import BaseAgent
 from ..utils import messaging, logging as insight_logging
 from ..utils.logging import Ledger
 from ..utils.retry import with_retry
-from ..utils.tracing import span
+from alpha_factory_v1.core.utils.tracing import span
 from typing import Callable, Awaitable, cast
 
 log = insight_logging.logging.getLogger(__name__)
