+++ pyproject.toml
@@ -20,6 +20,11 @@ packages = [
     "alpha_factory_v1.core",
     "alpha_factory_v1.core.interface",
     "alpha_factory_v1.core.interface.web_client",
+    "alpha_factory_v1.core.agents",
+    "alpha_factory_v1.core.evaluators",
+    "alpha_factory_v1.core.self_evolution",
+    "alpha_factory_v1.core.simulation",
+    "alpha_factory_v1.core.utils",
     "alpha_factory_v1.utils",
     "af_requests",
     "alpha_factory_v1.demos.meta_agentic_tree_search_v0",
