+++ tests/test_alert_webhook.py
@@ -7,9 +7,9 @@ import contextlib
 import logging
 import pytest
 
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src import orchestrator
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src.utils import alerts, messaging, config
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src.agents.base_agent import BaseAgent
+from alpha_factory_v1.core import orchestrator
+from alpha_factory_v1.core.utils import alerts, messaging, config
+from alpha_factory_v1.core.agents.base_agent import BaseAgent
 
 
 class DummyLedger:
