+++ tests/test_forecast.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 import math
 import pytest
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src.simulation import forecast, sector
+from alpha_factory_v1.core.simulation import forecast, sector
 from alpha_factory_v1.demos.meta_agentic_agi_v3.core.physics import gibbs
