+++ tests/test_multi_objective.py
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 import time
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src.simulation import mats
+from alpha_factory_v1.core.simulation import mats
 
 
 def _fit(acc: float, nov: float, lat: float) -> tuple[float, float, float]:
