+++ tests/test_orchestrator_backoff.py
@@ -2,8 +2,8 @@
 import asyncio
 import contextlib
 
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src import orchestrator
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src.utils import config
+from alpha_factory_v1.core import orchestrator
+from alpha_factory_v1.core.utils import config
 
 
 class FailingAgent(orchestrator.BaseAgent):
