+++ tests/test_orchestrator_bus_tls_env.py
@@ -21,8 +21,8 @@ try:
 except Exception:  # pragma: no cover - optional
     HAVE_CRYPTO = False
 
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src import orchestrator
-from alpha_factory_v1.demos.alpha_agi_insight_v1.src.utils import config, messaging
+from alpha_factory_v1.core import orchestrator
+from alpha_factory_v1.core.utils import config, messaging
 
 
 def _free_port() -> int:
